requirements.txt
# Core — LangChain 1.x + LangGraph 1.x (stable, released Nov 2025)
langchain>=1.0.0
langgraph>=1.0.0
langchain-openai>=0.3.0

# Google Calendar integration
langchain-google-community[calendar]>=3.0.0,<4.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Utilities
pytz>=2024.1
